Fix compilation warnings and enable Werror on gcc and clang (#4220)
* - clean clang warnings on mac local
- unskip -Werror on linux platforms
- fix cmake -Werror setting; add more fixes

* update

* fix

* clean

* change werror default to be false; update ci

* fix new compilation errors

* bump version to fix the generated tinysnb version mismatch

+++ tools/shell/embedded_shell.cpp
@@ -901,7 +901,7 @@ void EmbeddedShell::printExecutionResult(QueryResult& queryResult) const {
         if (queryResult.getNumColumns() == 1) {
             printf("(1 column)\n");
         } else {
-            printf("(%" PRIu64 " columns)\n", queryResult.getNumColumns());
+            printf("(%" PRIu64 " columns)\n", static_cast<uint64_t>(queryResult.getNumColumns()));
         }
         printf("Time: %.2fms (compiling), %.2fms (executing)\n", querySummary->getCompilingTime(),
             querySummary->getExecutionTime());
